Fix newlines added to chat box after every macro
The chat box has an invisible newline when empty. If you then set the
chat box to a newline, it adds an extra. So basically each time you run
a macro it adds another newline to the chat box (assuming you haven't
emptied it in the meantime). This is weird, but easily remedied with a
trim.