Intercept Content Security Policy headers
This allows the extension to work on Chrome, which seems to care a bit
more about security than my Nightly Firefox does at the moment.

Also, now that we're rewriting the CSP headers, we can inject code
directly into the script tags and not have to deal with creating blob
URLs.