Simplify roll API
Now none of the roll expression types are exposed to the user. The only
thing they deal with is RollExpr, which they create by calling
macroll.roll, macroll.num, or .plus/.minus/.multipliedBy/.dividedBy on a
RollExpr. This is more ergonomic than forcing the user to give a full
expression tree for each roll, and also still allows them to inspect a
roll using .inner.